+++ src/utils/queryUtils.ts
@@ -117,10 +117,12 @@ export function normalizeQuery(query: string): string {
       // Remove single line comments
       .replace(/^\/.+/gm, "")
       // Replace system commands
-      .replace(/^\\([a-zA-Z_1-2\\]+)[\t ]*(.*)/gm, (matched, command, args) =>
-        matched === "\\\\"
-          ? 'system"\\\\"'
-          : `system"${command} ${args.trim().replace(/"/gs, '\\"')}"`,
+      .replace(
+        /^\\([a-zA-Z_1-2\\]+(?::\d+)?)[\t ]*(.*)/gm,
+        (matched, command, args) =>
+          matched === "\\\\"
+            ? 'system"\\\\"'
+            : `system"${command} ${args.trim().replace(/"/gs, '\\"')}"`,
       )
       // Remove line comments
       .replace(
@@ -148,10 +150,12 @@ export function normalizeQSQLQuery(query: string): string {
       // Remove single line comments
       .replace(/^\/.+/gm, "")
       // Replace system commands
-      .replace(/^\\([a-zA-Z_1-2\\]+)[\t ]*(.*)/gm, (matched, command, args) =>
-        matched === "\\\\"
-          ? 'system"\\\\"'
-          : `system"${command} ${args.trim().replace(/"/gs, '\\"')}"`,
+      .replace(
+        /^\\([a-zA-Z_1-2\\]+(?::\d+)?)[\t ]*(.*)/gm,
+        (matched, command, args) =>
+          matched === "\\\\"
+            ? 'system"\\\\"'
+            : `system"${command} ${args.trim().replace(/"/gs, '\\"')}"`,
       )
       // Trim white space
       .trim()